refactor(auth): rename common folder within the components folder to auth refactor(index): update component exports to reflect new auth structure
