+++ client/src/components/index.ts
@@ -1,4 +1,3 @@
 export * from './header/Header';
-export * from './common/FormInput';
-export * from './common/SignInWithGoogle';
-export * from './common/AuthForm';
+export * from './auth/AuthForm';
+export * from './auth/SignInWithGoogle';
